chore: implement cli header, the corresponding code, and an example
