+++ CMakeLists.txt
@@ -47,7 +47,11 @@ add_library(commandershared
         ./deps/commander/commander.h
         ./deps/commander/commander.c
 )
-
+add_library(clishared
+    SHARED
+        ./include/utils/cli.h
+        ./src/utils/cli.c
+)
 add_library(loggershared
     SHARED
         ./include/utils/logger.h
@@ -60,6 +64,7 @@ target_compile_options(safememstatic PRIVATE ${flags})
 target_compile_options(randomshared PRIVATE ${flags})
 target_compile_options(safemathstatic PRIVATE ${flags})
 target_compile_options(commandershared PRIVATE ${flags})
+target_compile_options(clishared PRIVATE ${flags})
 
 # defines executables to build
 add_executable(cli-test ./src/cli.c)
@@ -78,10 +83,10 @@ target_compile_options(safe-math-test PRIVATE ${flags})
 target_compile_options(safe-mem-test PRIVATE ${flags})
 
 # links libraries
+target_link_libraries(cli-test clishared)
 target_link_libraries(cli-test safememstatic)
 target_link_libraries(cli-test loggershared)
 target_link_libraries(cli-test zlog)
-# target_link_libraries(cli-test argtable3)
 target_link_libraries(cli-test commandershared)
 target_link_libraries(logger-test cmocka)
 target_link_libraries(logger-test zlog)
