+++ include/utils/safe_mem.h
